pyproject.toml
[project]
name = "lpcv-2026"
version = "0.1.0"
description = "LPCVC 2026 Track 2: Video Classification with Dynamic Frame Selection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "av>=14.4.0",
    "datasets>=4.6.1",
    "evaluate>=0.4.6",
    "h5py>=3.16.0",
    "loguru>=0.7.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "torch>=2.10.0",
    "torchcodec>=0.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "typer>=0.24.1",
]


[project.scripts]
lpcv = "lpcv.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["lpcv"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "debugpy>=1.8.20",
    "ipython>=9.11.0",
    "pyright>=1.1.408",
    "ruff>=0.15.5",
]


[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = ["N812"]
